Faster q4_K and q5_K
Somehow, at least on my Ryzen CPU, _mm_hadd_epi16 seems to be
extremely slow. I was using _mm_hadd_epi16 the convert the
Q8_K sums for blocks of 16 to sums for blocks of 32 as needed
to compute the product with the 8 32-weight block scales of
q4_K/q5_K. Instead of doing that, I now shuffle the 8 q4_K/q5_K
block scales int 16 block scales to multiply directly with the
Q8_K block sums. This increases PP-512 performance by ~8% for
q4_K and q5_K. As a side benefit, the common code is now
factored out in struct Scales8K.